Perft Analysis Done
Couldn't do a rigorous perft analysis as the engine doesn't explore many
nodes per second as of now.
But as far as I've seen the board and move generation is now good
enough. But there are a few bugs that need to be fixed. A rigorous Perft
Analysis is very much essential.